+++ HMM1/src/data.jl
@@ -238,7 +238,7 @@ function loadHMM(fileName::String)
         A_ = zeros(N, N)
         for i in 1:N
             str = split(readline(io), ",")
-            A_[i,:] = parse.(Float32, str)
+            A_[i,:] = parse.(Float64, str)
         end
         A_=A(N, A_)
 
@@ -247,12 +247,12 @@ function loadHMM(fileName::String)
         B_ = zeros(N, M)
         for i in 1:N
             str = split(readline(io), ",")
-            B_[i,:] = parse.(Float32, str)
+            B_[i,:] = parse.(Float64, str)
         end
         B_ = B((N,M), B_)
 
         # Starting Distribution
-        startDist = parse.(Float32, split(readline(io), ",")) |> StochasticVector
+        startDist = parse.(Float64, split(readline(io), ",")) |> StochasticVector
 
         # Observation space
         observations = parse.(Int, split(readline(io), ",")) |> Set
