+++ HMM_Forecast/script/_analysis.jl
@@ -28,7 +28,7 @@ for state in states
 
     ## Generate Distro forecast
     println("-------------- Calc forecast distribution for {$N} states------------------")
-    distributionForecastVector = HMM_Forecast.createSeveralOneStepPredictions(hmm, dataTrainingAsIndeces, dataTestAsIndeces)::Vector{Vector{Float32}};
+    distributionForecastVector = HMM_Forecast.createSeveralOneStepPredictions(hmm, dataTrainingAsIndeces, dataTestAsIndeces)::Vector{Vector{Float64}};
     prevTime = HMM_Forecast.printTimeAndResetTimeStamp(prevTime)
 
     println("-------------- Generate PIT plots for {$N} states------------------")
@@ -61,7 +61,7 @@ for state in collect(states)
 
     ## Generate Distro forecast
     println("-------------- Calc forecast distribution for {$N} states------------------")
-    distributionForecastVector = HMM_Forecast.createSeveralOneStepPredictions(hmm, dataTrainingAsIndeces, dataTestAsIndeces)::Vector{Vector{Float32}};
+    distributionForecastVector = HMM_Forecast.createSeveralOneStepPredictions(hmm, dataTrainingAsIndeces, dataTestAsIndeces)::Vector{Vector{Float64}};
     prevTime = HMM_Forecast.printTimeAndResetTimeStamp(prevTime)
 
     println("-------------- Generate PIT plots for {$N} states------------------")
@@ -94,7 +94,7 @@ for state in collect(states)
 
     ## Generate Distro forecast
     println("-------------- Calc forecast distribution for {$N} states------------------")
-    distributionForecastVector = HMM_Forecast.createSeveralOneStepPredictions(hmm, dataTrainingAsIndeces, dataTestAsIndeces)::Vector{Vector{Float32}};
+    distributionForecastVector = HMM_Forecast.createSeveralOneStepPredictions(hmm, dataTrainingAsIndeces, dataTestAsIndeces)::Vector{Vector{Float64}};
     prevTime = HMM_Forecast.printTimeAndResetTimeStamp(prevTime)
 
     println("-------------- Generate PIT plots for {$N} states------------------")
@@ -127,7 +127,7 @@ for state in collect(states)
 
     ## Generate Distro forecast
     println("-------------- Calc forecast distribution for {$N} states------------------")
-    distributionForecastVector = HMM_Forecast.createSeveralOneStepPredictions(hmm, dataTrainingAsIndeces, dataTestAsIndeces)::Vector{Vector{Float32}};
+    distributionForecastVector = HMM_Forecast.createSeveralOneStepPredictions(hmm, dataTrainingAsIndeces, dataTestAsIndeces)::Vector{Vector{Float64}};
     prevTime = HMM_Forecast.printTimeAndResetTimeStamp(prevTime)
 
     println("-------------- Generate PIT plots for {$N} states------------------")
